apps/listings-service/requirements.txt
uvicorn[standard]
python-dotenv
tf2-utilities
aiohttp
fastapi
motor